extract profile photo url generation to service
